+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
+
 ## [0.0.1] - 2021-08-14
 
 ### This is the first release of *Unity Package com.pika.mekcoroutines*.
 
 - Migrated the custom version of mek from inside of unity.
+
+## [0.0.2] - 2022-08-11
+
+### Custom CoroutineWorkerEditor
+
+- Custom editor for CoroutineWorker.